waicy-checker-starter/requirements.txt
flask
flask-cors
python-dotenv
pandas
openpyxl
scikit-learn
